Adding global .jslint.json config file
ADAPTED FROM https://github.com/adobe/brackets/pull/4380/

If present, on project load or reload, a .jslint.json file in the project root will be loaded as the options passed to JSLint instead of the default brackets.json::jslint object.

Also setting JSLint as the JSON linter since it can handle that format as well.

TODO: Use DocumentManager to refresh the options if the .jslint.json file changes.