Conditioning .jslint.json loading based on CodeInspection.toggleEnabled
The command to enable or disable linting on file save is now managed by CodeInspection.js -- JSLint/main.js now checks the status of CodeInspection.toggleEnabled on load to see if it should load the project-wide .jslint.json file. It also adds a listener to the Command via CommandManager so it knows when to enable/disable the listeners on DocumentManager and ProjectManager to check for .jslint.json file changes/reloading.

This should fully re-implement the original project-wide JSLint config idea by @busykai which was proposed as a pull request before the linting system was abstracted into CodeInspection.js